Arm64 was built success on Android Studio
