+++ common/os_arch_features.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 #include "pac_kit.h"
 
-#include "PlatformUnifiedInterface/platform.h"
+#include "PlatformUnifiedInterface/platform_memory.h"
 
 #ifndef __has_feature
 #define __has_feature(x) 0
