+++ source/PlatformUnifiedInterface/platform.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "dobby/common.h"
+#include "platform_memory.h"
 
 namespace base {
 
@@ -56,31 +57,6 @@ struct OSThread : base::ThreadInterface, base::ThreadInterface::Delegate {
   OSThread(const char *name, uint32_t stack_size);
 
   bool Start();
-};
-
-enum MemoryPermission {
-  kNoAccess,
-  kRead = 1,
-  kWrite = 2,
-  kExecute = 4,
-  kReadWrite = kRead | kWrite,
-  kReadExecute = kRead | kExecute,
-  kReadWriteExecute = kRead | kWrite | kExecute,
-};
-
-class OSMemory {
-public:
-  static int PageSize();
-
-  static void *Allocate(size_t size, MemoryPermission access);
-
-  static void *Allocate(size_t size, MemoryPermission access, void *fixed_addr);
-
-  static bool Free(void *address, size_t size);
-
-  static bool Release(void *address, size_t size);
-
-  static bool SetPermission(void *address, size_t size, MemoryPermission access);
 };
 
 class OSPrint {
