+++ source/TrampolineBridge/ClosureTrampolineBridge/arm64/closure_bridge_arm64.S
@@ -50,8 +50,13 @@ str TMP_REG_0, [sp, #(1 * 8)]
 // call convention: x0 = register context, x1 = interceptor entry
 mov x0, sp
 ldr x1, [sp, #(2 * 8 + 2 * 8 + 30 * 8 + 8 * 16)]
+#if defined(__ANDROID__)
+adrp TMP_REG_0, :got:common_closure_bridge_handler
+ldr  TMP_REG_0, [TMP_REG_0, :got_lo12:common_closure_bridge_handler]
+#else
 adrp TMP_REG_0, cdecl(common_closure_bridge_handler)@PAGE
 add TMP_REG_0, TMP_REG_0, cdecl(common_closure_bridge_handler)@PAGEOFF
+#endif
 blr TMP_REG_0
 
 // restore stack, saved original sp
